reorganize new folder tests and add default empty folder test
